Update uploadClever.py
Added logic to eliminate those with the permission role of "Other" from the sections file since these are usually for administrative convenience rather than teaching. The list of exceptions to this is small and noted in the variable 'multiTeacher' to skip this new logic. Also added parents to the upload with logic to account for students who do not have parents at this time such as the one whose father died suddenly last week.